feat: 'The Surface' Landing Page - Portfolio Theme A
The landing page isn't finished yet. It is a form of see the portfolio info, for people that can't enable JS or for people that want to see quickly the basic info of the portfolio. This part of the portfolio has 'About Me', 'Projects', 'Experience', 'Certifications' and 'Contact' sections, for see 'Academy', 'Blog' sections or 'Contact' section with its all features, the user have to go to 'The System' (the most complete part of the portfolio).